Add start of test suite
This adds some tests that function by recording from fetching directly from the web, but then caches the return files so that when the tests are run next time, they simply use the cached HTTP responses. This allows us to mock the web service to avoid tests starting to fail on changes to the web responses. It would be nice to also have a suite that tests those and those will come later.

Whether or not HTTP responses are saved or not is controlled by the `record` flag at the top of the tests.